[BUGFIX] Allow vite to serve hidden files
The "deny all" snippet was copied from the default nginx configuration to the reverse proxy configuration for vite. However, vite needs to serve files from node_modules/.vite/* to work properly.

As this add-on is purely a local developer add-on and not expected to be run on production systems, it is not critical to serve dotfiles.